fix TCE recomendations
