+++ src/components/Categories.js
@@ -1,0 +1,31 @@
+import React from 'react';
+
+const Categories = () => {
+  const QuotesData = {
+    name: 'UNDER',
+    quote:
+            'CONSTRUCTION',
+  };
+
+  const Category = () => {
+    const { name, quote } = QuotesData;
+    return (
+      <>
+        <div className="content">
+          <div className="nameQuote">
+            {name}
+          </div>
+          <div className="quote">
+            {quote}
+          </div>
+        </div>
+      </>
+    );
+  };
+
+  return (
+    <Category />
+  );
+};
+
+export default Categories;
